+++ Assets/com.alelievr.NodeGraphProcessor/Runtime/Elements/RelayNode.cs
@@ -62,6 +62,9 @@ public class RelayNode : BaseNode
 	[CustomPortOutput(nameof(output), typeof(object), true)]
 	public void PushOutputs(List< SerializableEdge > edges, NodePort outputPort)
 	{
+		if (inputPorts.Count == 0)
+			return;
+
 		var inputPortEdges = inputPorts[0].GetEdges();
 
 		if (outputPort.portData.identifier != packIdentifier && outputIndex >= 0 && (unpackOutput || inputPortEdges.Count == 1))
@@ -115,6 +118,9 @@ public class RelayNode : BaseNode
 	[CustomPortBehavior(nameof(output))]
 	IEnumerable< PortData > OutputPortBehavior(List< SerializableEdge > edges)
 	{
+		if (inputPorts.Count == 0)
+			yield break;
+
 		var inputPortEdges = inputPorts[0].GetEdges();
 		var underlyingPortData = GetUnderlyingPortDataList();
 		if (unpackOutput && inputPortEdges.Count == 1)