+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Fixed relay node packing workflow and some deletion issues
+- Fixed ArgumentOutOfRange in relay node
 
 ## [0.8.0]